+++ playwright.config.js
@@ -1,6 +1,9 @@
 /** @type {import('@playwright/test').PlaywrightTestConfig} */
 
 import { defineConfig, devices } from '@playwright/test';
+import dotenv from 'dotenv';
+
+dotenv.config({ path: '.env.development' });
 
 export default defineConfig({
 	testDir: 'tests',
@@ -15,11 +18,33 @@ export default defineConfig({
 				storageState: 'tests/.auth/user.json'
 			},
 			dependencies: ['setup']
+		},
+		{
+			name: 'firefox',
+			use: {
+				...devices['Desktop Firefox'],
+				storageState: 'tests/.auth/user.json'
+			},
+			dependencies: ['setup']
 		}
 	],
 
-	webServer: {
-		command: 'npm run build && npm run preview',
-		port: 4173
+	webServer: [
+		{
+			command: './tests/start-test-server.sh 1.3.12',
+			port: 8000,
+			waitForPort: true,
+			stdout: 'pipe',
+			reuseExistingServer: !process.env.CI
+		},
+		{
+			command: 'npm run build && ORIGIN=http://localhost:5173 PORT=5173 node build',
+			port: 5173,
+			stdout: 'pipe',
+			reuseExistingServer: !process.env.CI
+		}
+	],
+	use: {
+		baseURL: 'http://localhost:5173'
 	}
 });