+++ tests/home.spec.js
@@ -2,5 +2,5 @@ import { expect, test } from '@playwright/test';
 
 test('index page has expected h1', async ({ page }) => {
 	await page.goto('/');
-	expect(await page.textContent('h1')).toBe('Fractal web client');
+	expect(await page.textContent('h1')).toBe('Welcome to Fractal web client.');
 });