+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 *Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web repository.*
 
+# Unreleased
+
+* Added end to end tests and coverage configuration (\#320).
+
 # 0.6.1
 
 This release works best with fractal-server >=1.3.12 (since it makes use of the